Speed up common case of decoding
Turns out that one of our internal services at work was spending
an unnecessary 10% of its time on URI decoding (really! It is a
microservice). With this change, that should drop to at most 3%.
In a nutshell, it eschews calling the full strtol where that
is not necessary and uses a custom converter instead.